+++ packages/flutter_cockpit_test/lib/flutter_cockpit_test.dart
@@ -1,0 +1,5 @@
+library;
+
+export 'src/cockpit_test.dart';
+export 'src/cockpit_test_options.dart';
+export 'src/cockpit_native_tester.dart';